My first Android application developed using Java.
This is a basic converter that has a menu with different conversions supported.
There is also a support for a custom conversion that can be set by the user, by providing the names and the formula.

The scope of creating this was to get used to AndroidStudio IDE and Java.